requirements.txt
azure-functions
azure-storage-blob
azure-cosmosdb-table